Fix asar loading which was broken due to node bootstrap refactoring
