+++ lib/internal/bootstrap/loaders.js
@@ -50,7 +50,9 @@
     enumerable: true,
     writable: false
   });
+  process.NativeModule = NativeModule;
 
+  var localGlobal = global
   // Set up process.binding() and process._linkedBinding()
   {
     const bindingObj = Object.create(null);
@@ -99,7 +101,7 @@
     this.loading = false;
   }
 
-  NativeModule._source = getBinding('natives');
+  NativeModule._source = process.binding('natives');
   NativeModule._cache = {};
 
   const config = getBinding('config');
@@ -189,9 +191,10 @@
   };
 
   NativeModule.wrapper = [
-    '(function (exports, require, module, process) {',
-    '\n});'
-  ];
+    '(function (exports, require, module, process, global, Buffer) { ' +
+    'return function (exports, require, module, process) { ',
+    '\n}.call(this, exports, require, module, process); });'
+  ]
 
   NativeModule.prototype.compile = function() {
     let source = NativeModule.getSource(this.id);
@@ -206,7 +209,7 @@
       const requireFn = this.id.startsWith('internal/deps/') ?
         NativeModule.requireForDeps :
         NativeModule.require;
-      fn(this.exports, requireFn, this, process);
+      fn(this.exports, requireFn, this, process, localGlobal, localGlobal.Buffer);
 
       this.loaded = true;
     } finally {
