+++ lib/internal/bootstrap/node.js
@@ -24,6 +24,8 @@
 
     EventEmitter.call(process);
 
+    process.NativeModule = NativeModule;
+
     setupProcessObject();
 
     // do this good and early, since it handles errors.
@@ -59,6 +61,7 @@
     _process.setup_performance();
     _process.setup_cpuUsage();
     _process.setupMemoryUsage();
+    setupAsarSupport();
     _process.setupKillAndExit();
     if (global.__coverage__)
       NativeModule.require('internal/process/write-coverage').setup();
@@ -570,6 +573,10 @@
     source = CJSModule.wrap(source);
     // compile the script, this will throw if it fails
     new vm.Script(source, { displayErrors: true, filename });
+  }
+
+  function setupAsarSupport() {
+    process.binding('atom_common_asar').initAsarSupport(process, NativeModule.require);
   }
 
   startup();
